python3Packages.retworkx: convert to importCargo
This doesn't use exactly the same lockfile as in the patch that was
used previously, because I think the cargo that generated it had a
bug.  It didn't include every version of every dependency, which meant
that importCargo couldn't work.  But just running cargo check produced
a fixed version, so this patch uses that instead.