stdenv: inherit most hostPlatform is* attrs
We were previously very inconsistent about which of these we inherited,
so let's inherit all of them (except isCompatible, which is a different
thing, being a function).  Otherwise, it can be quite confusing to see
stdenv.is* checks next to stdenv.hostPlatform.is* checks, which leads
one to wonder what the difference is (there isn't one).