bash_5: fix build
Even though we were patching both configure.ac and the generated
configure script, bash still wanted to rerun autoconf.  The easiest
thing to do here is just to let it, which means we can remove the bit
of the patch that modified the generated script, since it will now be
regenerated.

Fixes: https://github.com/NixOS/nixpkgs/issues/111805