pkgsCross.wasm32,pkgsCross.wasm64: init
I've called the doubles e.g. "wasm32-none", but the triples
e.g. "wasm32-unknown-unknown".  The reason for the kernel discrepancy
is that `config.sub wasm32` outputs "wasm32-unknown-none", so a
hypothetical Nix compiled for wasm32 would call its system
"wasm32-none", but LLVM uses "unknown" for the kernel.  So we
compromise.