nixosTests.virtualbox: remove
As far as I can tell, this test has been broken since 2019.  I've
spent the last couple of weeks trying to resurrect it, but wasn't able
to.  One significant problem is that at some point the
useKvmNestedVirt option broke, but since Virtualbox 6.1 software
virtualisation isn't supported, so using nested hardware
virtualisation would be the only way to run the test.

Anyway, with it having been broken for years, we're clearly not
getting any value out of keeping it around, so I think it's time to
remove it.